+++ apps/web/src/pages/JobPosterTemplates/JobPosterTemplatePage/components/AssetTechnicalSkills.tsx
