+++ apps/web/src/pages/JobPosterTemplates/JobPosterTemplatePage/components/BasicDetails.tsx
