+++ apps/web/src/pages/JobPosterTemplates/JobPosterTemplatePage/components/EssentialBehaviouralSkills.tsx
