+++ apps/web/src/pages/JobPosterTemplates/JobPosterTemplatePage/components/EssentialTechnicalSkills.tsx
