+++ apps/web/src/pages/JobPosterTemplates/JobPosterTemplatePage/components/KeyTasks.tsx
