+++ apps/web/src/pages/JobPosterTemplates/JobPosterTemplatePage/templateSkills.ts
